tests/AutomationScripts/ui_automation/requirements.txt
pytest>=8.0.0
playwright>=1.52.0
pytest-playwright>=0.7.0
pytest-html>=4.1.1
PyYAML>=6.0.1
allure-pytest>=2.13.5
